# Optimize `mapMonotonic` by using an `Fmap` class instead of `Functor`